Updated for v3.5.1

+++ Vagrantfile
@@ -1,4 +1,4 @@
-# Created by Topology-Converter v4.6.5
+# Created by Topology-Converter v4.6.6
 #    Template Revision: v4.6.5
 #    https://github.com/cumulusnetworks/topology_converter
 #    using topology data from: ./topology.dot
@@ -11,54 +11,60 @@
 
 
 
+Vagrant.require_version ">= 1.8.6"
 
 $script = <<-SCRIPT
 if grep -q -i 'cumulus' /etc/lsb-release &> /dev/null; then
     echo "### RUNNING CUMULUS EXTRA CONFIG ###"
     source /etc/lsb-release
-    if [[ $DISTRIB_RELEASE =~ ^2.* ]]; then
-        echo "  INFO: Detected a 2.5.x Based Release"
+    if [ -z /etc/app-release ]; then
+        echo "  INFO: Detected NetQ TS Server"
+        source /etc/app-release
+        echo "  INFO: Running NetQ TS Appliance Version $APPLIANCE_VERSION"
+    else
+        if [[ $DISTRIB_RELEASE =~ ^2.* ]]; then
+            echo "  INFO: Detected a 2.5.x Based Release"
 
-        echo "  adding fake cl-acltool..."
-        echo -e "#!/bin/bash\nexit 0" > /usr/bin/cl-acltool
-        chmod 755 /usr/bin/cl-acltool
+            echo "  adding fake cl-acltool..."
+            echo -e "#!/bin/bash\nexit 0" > /usr/bin/cl-acltool
+            chmod 755 /usr/bin/cl-acltool
 
-        echo "  adding fake cl-license..."
-        echo -e "#!/bin/bash\nexit 0" > /usr/bin/cl-license
-        chmod 755 /usr/bin/cl-license
+            echo "  adding fake cl-license..."
+            echo -e "#!/bin/bash\nexit 0" > /usr/bin/cl-license
+            chmod 755 /usr/bin/cl-license
 
-        echo "  Disabling default remap on Cumulus VX..."
-        mv -v /etc/init.d/rename_eth_swp /etc/init.d/rename_eth_swp.backup
+            echo "  Disabling default remap on Cumulus VX..."
+            mv -v /etc/init.d/rename_eth_swp /etc/init.d/rename_eth_swp.backup
 
-        echo "### Rebooting to Apply Remap..."
+            echo "### Rebooting to Apply Remap..."
 
-    elif [[ $DISTRIB_RELEASE =~ ^3.* ]]; then
-        echo "  INFO: Detected a 3.x Based Release"
-        echo "### Disabling default remap on Cumulus VX..."
-        mv -v /etc/hw_init.d/S10rename_eth_swp.sh /etc/S10rename_eth_swp.sh.backup &> /dev/null
-        echo "### Disabling ZTP service..."
-        systemctl stop ztp.service
-        ztp -d 2>&1
-        echo "### Resetting ZTP to work next boot..."
-        ztp -R 2>&1
-        echo "  INFO: Detected Cumulus Linux v$DISTRIB_RELEASE Release"
-        if [[ $DISTRIB_RELEASE =~ ^3.[1-9].* ]]; then
-            echo "### Fixing ONIE DHCP to avoid Vagrant Interface ###"
-            echo "     Note: Installing from ONIE will undo these changes." 
-            mkdir /tmp/foo
-            mount LABEL=ONIE-BOOT /tmp/foo
-            sed -i 's/eth0/eth1/g' /tmp/foo/grub/grub.cfg
-            sed -i 's/eth0/eth1/g' /tmp/foo/onie/grub/grub-extra.cfg
-            umount /tmp/foo
-        fi
-        if [[ $DISTRIB_RELEASE =~ ^3.[2-9].* ]]; then
-            if [[ $(grep "vagrant" /etc/netd.conf | wc -l ) == 0 ]]; then
-                echo "### Giving Vagrant User Ability to Run NCLU Commands ###"
-                sed -i 's/users_with_edit = root, cumulus/users_with_edit = root, cumulus, vagrant/g' /etc/netd.conf
-                sed -i 's/users_with_show = root, cumulus/users_with_show = root, cumulus, vagrant/g' /etc/netd.conf
+        elif [[ $DISTRIB_RELEASE =~ ^3.* ]]; then
+            echo "  INFO: Detected a 3.x Based Release"
+            echo "### Disabling default remap on Cumulus VX..."
+            mv -v /etc/hw_init.d/S10rename_eth_swp.sh /etc/S10rename_eth_swp.sh.backup &> /dev/null
+            echo "### Disabling ZTP service..."
+            systemctl stop ztp.service
+            ztp -d 2>&1
+            echo "### Resetting ZTP to work next boot..."
+            ztp -R 2>&1
+            echo "  INFO: Detected Cumulus Linux v$DISTRIB_RELEASE Release"
+            if [[ $DISTRIB_RELEASE =~ ^3.[1-9].* ]]; then
+                echo "### Fixing ONIE DHCP to avoid Vagrant Interface ###"
+                echo "     Note: Installing from ONIE will undo these changes." 
+                mkdir /tmp/foo
+                mount LABEL=ONIE-BOOT /tmp/foo
+                sed -i 's/eth0/eth1/g' /tmp/foo/grub/grub.cfg
+                sed -i 's/eth0/eth1/g' /tmp/foo/onie/grub/grub-extra.cfg
+                umount /tmp/foo
+            fi
+            if [[ $DISTRIB_RELEASE =~ ^3.[2-9].* ]]; then
+                if [[ $(grep "vagrant" /etc/netd.conf | wc -l ) == 0 ]]; then
+                    echo "### Giving Vagrant User Ability to Run NCLU Commands ###"
+                    sed -i 's/users_with_edit = root, cumulus/users_with_edit = root, cumulus, vagrant/g' /etc/netd.conf
+                    sed -i 's/users_with_show = root, cumulus/users_with_show = root, cumulus, vagrant/g' /etc/netd.conf
+                fi
             fi
         fi
-
     fi
 fi
 echo "### DONE ###"
@@ -68,7 +74,7 @@ SCRIPT
 
 Vagrant.configure("2") do |config|
 
-  simid = 1515009481
+  simid = 1516247777
 
   config.vm.provider "virtualbox" do |v|
     v.gui=false
@@ -146,7 +152,7 @@ end
     device.vm.hostname = "oob-mgmt-switch" 
     
     device.vm.box = "CumulusCommunity/cumulus-vx"
-    device.vm.box_version = "3.5.0"
+    device.vm.box_version = "3.5.1"
     device.vm.provider "virtualbox" do |v|
       v.name = "#{simid}_oob-mgmt-switch"
       v.customize ["modifyvm", :id, '--audiocontroller', 'AC97', '--audio', 'Null']
@@ -320,7 +326,7 @@ end
     device.vm.hostname = "exit02" 
     
     device.vm.box = "CumulusCommunity/cumulus-vx"
-    device.vm.box_version = "3.5.0"
+    device.vm.box_version = "3.5.1"
     device.vm.provider "virtualbox" do |v|
       v.name = "#{simid}_exit02"
       v.customize ["modifyvm", :id, '--audiocontroller', 'AC97', '--audio', 'Null']
@@ -462,7 +468,7 @@ end
     device.vm.hostname = "exit01" 
     
     device.vm.box = "CumulusCommunity/cumulus-vx"
-    device.vm.box_version = "3.5.0"
+    device.vm.box_version = "3.5.1"
     device.vm.provider "virtualbox" do |v|
       v.name = "#{simid}_exit01"
       v.customize ["modifyvm", :id, '--audiocontroller', 'AC97', '--audio', 'Null']
@@ -604,7 +610,7 @@ end
     device.vm.hostname = "spine02" 
     
     device.vm.box = "CumulusCommunity/cumulus-vx"
-    device.vm.box_version = "3.5.0"
+    device.vm.box_version = "3.5.1"
     device.vm.provider "virtualbox" do |v|
       v.name = "#{simid}_spine02"
       v.customize ["modifyvm", :id, '--audiocontroller', 'AC97', '--audio', 'Null']
@@ -730,7 +736,7 @@ end
     device.vm.hostname = "spine01" 
     
     device.vm.box = "CumulusCommunity/cumulus-vx"
-    device.vm.box_version = "3.5.0"
+    device.vm.box_version = "3.5.1"
     device.vm.provider "virtualbox" do |v|
       v.name = "#{simid}_spine01"
       v.customize ["modifyvm", :id, '--audiocontroller', 'AC97', '--audio', 'Null']
@@ -856,7 +862,7 @@ end
     device.vm.hostname = "leaf04" 
     
     device.vm.box = "CumulusCommunity/cumulus-vx"
-    device.vm.box_version = "3.5.0"
+    device.vm.box_version = "3.5.1"
     device.vm.provider "virtualbox" do |v|
       v.name = "#{simid}_leaf04"
       v.customize ["modifyvm", :id, '--audiocontroller', 'AC97', '--audio', 'Null']
@@ -998,7 +1004,7 @@ end
     device.vm.hostname = "leaf02" 
     
     device.vm.box = "CumulusCommunity/cumulus-vx"
-    device.vm.box_version = "3.5.0"
+    device.vm.box_version = "3.5.1"
     device.vm.provider "virtualbox" do |v|
       v.name = "#{simid}_leaf02"
       v.customize ["modifyvm", :id, '--audiocontroller', 'AC97', '--audio', 'Null']
@@ -1140,7 +1146,7 @@ end
     device.vm.hostname = "leaf03" 
     
     device.vm.box = "CumulusCommunity/cumulus-vx"
-    device.vm.box_version = "3.5.0"
+    device.vm.box_version = "3.5.1"
     device.vm.provider "virtualbox" do |v|
       v.name = "#{simid}_leaf03"
       v.customize ["modifyvm", :id, '--audiocontroller', 'AC97', '--audio', 'Null']
@@ -1282,7 +1288,7 @@ end
     device.vm.hostname = "leaf01" 
     
     device.vm.box = "CumulusCommunity/cumulus-vx"
-    device.vm.box_version = "3.5.0"
+    device.vm.box_version = "3.5.1"
     device.vm.provider "virtualbox" do |v|
       v.name = "#{simid}_leaf01"
       v.customize ["modifyvm", :id, '--audiocontroller', 'AC97', '--audio', 'Null']
@@ -1824,7 +1830,7 @@ end
     device.vm.hostname = "internet" 
     
     device.vm.box = "CumulusCommunity/cumulus-vx"
-    device.vm.box_version = "3.5.0"
+    device.vm.box_version = "3.5.1"
     device.vm.provider "virtualbox" do |v|
       v.name = "#{simid}_internet"
       v.customize ["modifyvm", :id, '--audiocontroller', 'AC97', '--audio', 'Null']
